Supabase_cloud-dashboard/requirements.txt
streamlit==1.52.1
plotly==6.0.1
pandas==2.1.4
numpy==1.26.4
psycopg2-binary==2.9.11
